When converting the measumrent rate from a double to a string, use the same formatting string as Ascension.  This time it's really, really right. :)
git-svn-id: https://vrjuggler.googlecode.com/svn/branches/RELENG_1_0/juggler_1.0@4243 08b38cba-cd3b-11de-854e-f91c5b6e4272

+++ Input/vjPosition/aMotionStar.cpp
@@ -1627,7 +1627,7 @@ aMotionStar::convertMeasurementRate (const double rate, std::string& str_rate)
 {
     char rate_a[7];
 
-    snprintf(rate_a, 7, "%d", (int) (rate * 1000.0));
+    snprintf(rate_a, 7, "%06.0f", rate * 1000.0);
     str_rate = rate_a;
 }
